Sensor: README.md: Add repository information
It may be a good idea to display the repository content information in
all of the test result branches, in case some link does not lead to the
'master' -branch.

On the other side, this may also be a bad idea, as if new branches are
added, all updates to the README need to be carried over to all of the
branches. So, as with all maintenance branches, this scales well only
when amount of updates or branches stays low.

Currently both of those are low. Hence the benefit of the availability
of the information currently outweighs the burden of the updates.

Add README contents to the Sensor branch too.